+++ interface/src/main/java/com/aws/peach/interfaces/api/OrderController.java
@@ -1,6 +1,5 @@
 package com.aws.peach.interfaces.api;
 
-import com.aws.peach.application.order.OrderViewService;
 import com.aws.peach.application.order.PlaceOrderService;
 import com.aws.peach.domain.order.entity.Order;
 import org.springframework.http.ResponseEntity;
